+++ src/drf_yasg/inspectors/field.py
@@ -429,7 +429,6 @@ else:
         """Hack to force ``djangorestframework_camel_case`` to camelize a plain string."""
         return next(iter(camelize({s: ''})))
 
-
     def camelize_schema(schema_or_ref, components):
         """Recursively camelize property names for the given schema using ``djangorestframework_camel_case``."""
         schema = openapi.resolve_ref(schema_or_ref, components)
@@ -443,7 +442,6 @@ else:
                 schema.required = [camelize_string(p) for p in schema.required]
 
         return schema_or_ref
-
 
     class CamelCaseJSONFilter(FieldInspector):
         def is_camel_case(self):