+++ testproj/createsuperuser.py
@@ -1,0 +1,4 @@
+from django.contrib.auth.models import User
+
+User.objects.filter(username='admin').delete()
+User.objects.create_superuser('admin', 'admin@admin.admin', 'passwordadmin')