+++ testproj/snippets/views.py
@@ -1,17 +1,27 @@
 from djangorestframework_camel_case.parser import CamelCaseJSONParser
 from djangorestframework_camel_case.render import CamelCaseJSONRenderer
+from inflection import camelize
 from rest_framework import generics
 
+from drf_yasg.inspectors import SwaggerAutoSchema
 from snippets.models import Snippet
 from snippets.serializers import SnippetSerializer
+
+
+class CamelCaseOperationIDAutoSchema(SwaggerAutoSchema):
+    def get_operation_id(self, operation_keys):
+        operation_id = super(CamelCaseOperationIDAutoSchema, self).get_operation_id(operation_keys)
+        return camelize(operation_id, uppercase_first_letter=False)
 
 
 class SnippetList(generics.ListCreateAPIView):
     """SnippetList classdoc"""
     queryset = Snippet.objects.all()
     serializer_class = SnippetSerializer
+
     parser_classes = (CamelCaseJSONParser,)
     renderer_classes = (CamelCaseJSONRenderer,)
+    swagger_schema = CamelCaseOperationIDAutoSchema
 
     def perform_create(self, serializer):
         serializer.save(owner=self.request.user)
@@ -34,8 +44,10 @@ class SnippetDetail(generics.RetrieveUpdateDestroyAPIView):
     queryset = Snippet.objects.all()
     serializer_class = SnippetSerializer
     pagination_class = None
+
     parser_classes = (CamelCaseJSONParser,)
     renderer_classes = (CamelCaseJSONRenderer,)
+    swagger_schema = CamelCaseOperationIDAutoSchema
 
     def patch(self, request, *args, **kwargs):
         """patch method docstring"""